Chore refactor header file structure
